Full Level Design: Box Life
Feel free to change anything you like as long as you feel that you are making it better than before. The only thing I ask that you don't change is the Shortcut path that connects to the PlayerStart area. This is meant for any player who gets greedy about picking up items along this route.